cmp_hints: widen LRU 'oldest' to uint64_t to match field
pool_add_locked() reads pool->entries[i].last_used (uint64_t) into
a local declared `unsigned int oldest`, which truncates the high
32 bits on every assignment.  Once pool->last_used_stamp crosses
2^32 the comparison wraps and the eviction policy starts picking
arbitrary victims instead of the genuinely-oldest entry.

The header field was deliberately widened to uint64_t to avoid
exactly this multi-day-run wrap; cmp_hints.h documents the intent
inline.  Match the local to the field so the widening fix can
actually take effect.

+++ cmp_hints.c
@@ -238,7 +238,7 @@ static bool pool_add_locked(struct cmp_hint_pool *pool,
 	unsigned int i, count = pool->count;
 	uint64_t stamp = ++pool->last_used_stamp;
 	unsigned int victim;
-	unsigned int oldest;
+	uint64_t oldest;
 
 	for (i = 0; i < count; i++) {
 		struct cmp_hint_entry *e = &pool->entries[i];
